+++ packages/eslint-config/test/fixtures/basic/eslint.config.mjs
@@ -1,0 +1,11 @@
+import {
+  defineFlatConfig,
+  imports,
+  javascript,
+  typescript,
+} from '../../../dist/index.mjs'
+
+export default defineFlatConfig([...imports, ...javascript, ...typescript], {
+  prettierEnabled: false,
+  tailwindcssEnabled: false,
+})